-fix: fixed all the dissapearance of my UI and the logics of NewCollegeForm.
